[UPDATE]changed components location
changed components location and little change of App.js
